+++ NEWS.md
@@ -1,4 +1,4 @@
-# rhino 1.1.1.9999
+# rhino (development version)
 
 * Added a `NEWS.md` file to track changes to the package.
-* Adds `paths` argument to `lint_r()`. `paths` defaults to `NULL` and should lint `app` and `tests/testthat` as default. `paths` should take a vector of directory or file paths and return lint error or success messages whichever is appropriate. 
+* Added a `paths` argument to `lint_r()`.